Moved auth UI components to another directory
