+++ composeApp/src/commonMain/kotlin/band/effective/office/elevator/components/auth_components/AuthTitle.kt
@@ -1,4 +1,4 @@
-package band.effective.office.elevator.components
+package band.effective.office.elevator.components.auth_components
 
 import androidx.compose.foundation.layout.wrapContentSize
 import androidx.compose.material.Text
@@ -8,10 +8,9 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.sp
-import band.effective.office.elevator.theme_light_tertiary_color
 
 @Composable
-fun AuthSubTitle(
+fun AuthTitle(
     text: String,
     modifier: Modifier = Modifier,
     textAlign: TextAlign?,
@@ -21,11 +20,11 @@ fun AuthSubTitle(
         modifier = Modifier
             .wrapContentSize()
             .then(modifier),
-        fontSize = 16.sp,
-//    fontFamily FontFamily TODO : Add font
-        fontWeight = FontWeight(500),
+        fontSize = 28.sp,
+//        fontFamily = FontFamily TODO : Add font
+        fontWeight = FontWeight(600),
+        color = Color(0xFF5800CB),
         textAlign = textAlign,
-        color = theme_light_tertiary_color,
         letterSpacing = 0.1.sp
     )
 }
